PDV-283 | resolved bugs
