+++ components/AddProduct/index.tsx
@@ -123,6 +123,7 @@ const AddProduct = ({ handleAlerts, updateList }: Props) => {
               <NumberInput
                 id="add-product-price"
                 label="Preço do produto *"
+                placeholder="R$ 50,00"
                 handleChange={handleChange}
                 value={values?.price || ''}
                 error={errorObject.price}
