+++ components/ProductItem/index.tsx
@@ -92,7 +92,7 @@ const ProductItem = ({
   };
 
   const updateImage = async () => {
-    const response = await fetchPutProduct({ images: [imgValue] });
+    const response = await fetchPutProduct({ images: imgValue !== '' ? [imgValue] : [] });
     !response.message && updateList();
     setOpenDialog(false);
     handleAlerts(response);
@@ -181,7 +181,7 @@ const ProductItem = ({
             />
           ) : (
             <Typography component="p">
-              {currencyFormatter(product.price)}
+              {product.price !== 0 ? currencyFormatter(product.price) : "R$ 0,00"}
             </Typography>
           )}
           {isAdmin && (
